@@ -19,11 +19,10 @@
     "path-browserify": "^1.0.1",
     "pinia": "^2.0.21",
     "subscriptions-transport-ws": "^0.11.0",
-    "vue": "^2.7.14",
-    "vue-class-component": "^7.2.3",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.6.4",
-    "vuetify": "^2.6.15"
+    "vue": "^3.3.4",
+    "@vue/compat": "^3.3.4",
+    "vue-router": "^4.2.3",
+    "vuetify": "^3.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
@@ -35,7 +34,7 @@
     "@types/node": "^18.15.11",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
-    "@vitejs/plugin-vue2": "^1.1.2",
+    "@vitejs/plugin-vue": "^4.2.3",
     "@vue/eslint-config-typescript": "^11.0.0",
     "eslint": "^8.22.0",
     "eslint-config-prettier": "^8.3.0",
@@ -47,7 +46,6 @@
     "sass": "^1.54.5",
     "ts-node": "^10.9.1",
     "typescript": "~5.1.6",
-    "unplugin-vue-components": "^0.22.4",
     "vite": "^4.3.9",
     "vite-plugin-monaco-editor": "^1.1.0",
     "vite-plugin-package-version": "^1.0.2"
