@@ -1,4 +1,4 @@
-import Vue from "vue";
+import { createApp, h } from "vue";
 
 import pinia from "@/stores";
 import router from "@/router";
@@ -7,14 +7,13 @@ import { defineThemes } from "@/monaco-editor";
 
 import App from "@/app/AppWLoadConfig.vue";
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  vuetify,
-  pinia,
+const app = createApp({
   created() {
     defineThemes(this.$vuetify);
   },
-  render: (h) => h(App),
-}).$mount("#app");
+  render: () => h(App),
+});
+app.use(pinia);
+app.use(router);
+app.use(vuetify);
+app.mount("#app");
