@@ -1,35 +1,41 @@
 <template>
-  <v-card flat height="100%" class="grey lighten-4">
-    <v-system-bar color="primary" dark>
-      <v-icon> mdi-language-python </v-icon>
-      <span>&lt;string&gt;</span>
-    </v-system-bar>
-    <div class="g-container">
-      <v-card-actions class="flex-row flex-wrap pa-1 grey lighten-4">
-        <v-tooltip bottom open-delay="500" v-for="(b, i) in buttons" :key="i">
-          <template v-slot:activator="{ on, attrs }">
-            <v-btn
-              color="primary"
-              icon
-              outlined
-              @click="onClick(b.method)"
-              :disabled="b.disabled"
-              v-bind="attrs"
-              v-on="on"
-              class="ma-1"
-            >
-              <v-icon>{{ b.icon }}</v-icon>
-            </v-btn>
-          </template>
-          <span>{{ b.text }}</span>
-        </v-tooltip>
-      </v-card-actions>
-      <v-divider></v-divider>
-      <div style="height: 100%">
-        <div ref="editor" style="height: 100%; max-height: 100%"></div>
+  <v-layout full-height style="width: 100%">
+    <!--
+    v-layout-can be removed when v-system-bar is removed.
+    v-layout is there to prevent the height of v-system-bar from being included
+    in the padding top of v-main.
+   -->
+    <v-card flat height="100%" class="bg-grey-lighten-4" style="width: 100%">
+      <v-system-bar color="primary" dark style="position: static">
+        <v-icon> mdi-language-python </v-icon>
+        <span>&lt;string&gt;</span>
+      </v-system-bar>
+      <div class="g-container">
+        <v-card-actions class="flex-row flex-wrap pa-1 grey-lighten-4">
+          <v-tooltip bottom open-delay="500" v-for="(b, i) in buttons" :key="i">
+            <template v-slot:activator="{ props }">
+              <v-btn
+                v-bind="props"
+                color="primary"
+                icon
+                variant="outlined"
+                @click="onClick(b.method)"
+                :disabled="b.disabled"
+                class="ma-1"
+              >
+                <v-icon>{{ b.icon }}</v-icon>
+              </v-btn>
+            </template>
+            <span>{{ b.text }}</span>
+          </v-tooltip>
+        </v-card-actions>
+        <v-divider></v-divider>
+        <div style="height: 100%">
+          <div ref="editor" style="height: 100%; max-height: 100%"></div>
+        </div>
       </div>
-    </div>
-  </v-card>
+    </v-card>
+  </v-layout>
 </template>
 
 <script setup lang="ts">
@@ -40,11 +46,11 @@ import * as monaco from "monaco-editor";
 import { useSourceQuery, useResetMutation } from "@/gql/graphql";
 
 interface Props {
-  value: boolean;
+  modelValue: boolean;
 }
 
 interface Emits {
-  (e: "input", value: boolean): void;
+  (event: "update:modelValue", value: boolean): void;
 }
 
 const props = defineProps<Props>();
@@ -113,7 +119,7 @@ const editing = computed(() => {
 });
 
 watch(editing, (val) => {
-  emit("input", val);
+  emit("update:modelValue", val);
 });
 
 type Method = "save" | "reset";
