@@ -1,32 +1,32 @@
 <template>
-  <v-card flat height="100%" color="background lighten-4">
-    <v-system-bar color="primary" dark>
-      <v-icon>mdi-window-maximize</v-icon>
-      <span>stdout</span>
-      <v-spacer></v-spacer>
-      <v-tooltip bottom open-delay="500">
-        <template v-slot:activator="{ on, attrs }">
-          <v-btn
-            icon
-            x-small
-            :disabled="!data?.stdout"
-            @click="clear"
-            v-bind="attrs"
-            v-on="on"
-          >
-            <v-icon>mdi-eraser</v-icon>
-          </v-btn>
-        </template>
-        <span>Clear</span>
-      </v-tooltip>
-    </v-system-bar>
-    <!-- <v-app-bar dense flat color="secondary"></v-app-bar> -->
-    <v-card-text style="height: calc(100% - 24px)" class="py-1">
-      <pre
-        style="height: 100%"
-        class="overflow-auto">{{ data?.stdout }}<span ref="bottom"></span></pre>
-    </v-card-text>
-  </v-card>
+  <v-layout full-height style="width: 100%">
+    <!--
+    v-layout-can be removed when v-system-bar is removed.
+    v-layout is there to prevent the height of v-system-bar from being included
+    in the padding top of v-main.
+   -->
+    <v-card flat height="100%" color="background-lighten-4" style="width: 100%">
+      <v-system-bar color="primary" dark style="position: static">
+        <v-icon>mdi-window-maximize</v-icon>
+        <span>stdout</span>
+        <v-spacer></v-spacer>
+        <v-tooltip bottom open-delay="500">
+          <template v-slot:activator="{ props }">
+            <v-icon v-bind="props" :disabled="!data?.stdout" @click="clear">
+              mdi-eraser
+            </v-icon>
+          </template>
+          <span>Clear</span>
+        </v-tooltip>
+      </v-system-bar>
+      <!-- <v-app-bar dense flat color="secondary"></v-app-bar> -->
+      <v-card-text style="height: calc(100% - 24px)" class="py-1">
+        <pre
+          style="height: 100%"
+          class="overflow-auto">{{ data?.stdout }}<span ref="bottom"></span></pre>
+      </v-card-text>
+    </v-card>
+  </v-layout>
 </template>
 
 <script setup lang="ts">
