@@ -1,4 +1,4 @@
-declare module "vuetify/lib/framework" {
+declare module "vuetify/framework" {
   import Vuetify from "vuetify";
   export default Vuetify;
 }
