@@ -1,93 +1,101 @@
 <template>
   <div class="g-container">
-    <v-card flat min-width="400px" color="background lighten-4">
-      <v-system-bar color="primary" dark>
-        <span> System bar </span>
-        <v-spacer></v-spacer>
-        <v-icon>mdi-minus</v-icon>
-        <v-icon>mdi-checkbox-blank-outline</v-icon>
-        <v-icon>mdi-close</v-icon>
-      </v-system-bar>
-      <v-toolbar flat color="primary lighten-2" dark extension-height="36px">
-        <v-app-bar-nav-icon></v-app-bar-nav-icon>
-        <v-toolbar-title> Tool bar </v-toolbar-title>
-        <v-spacer></v-spacer>
-        <v-btn icon>
-          <v-icon>mdi-magnify</v-icon>
-        </v-btn>
-        <template v-slot:extension>
-          <v-tabs
-            v-model="tab"
-            align-with-title
-            height="36px"
-            hide-slider
-            active-class="background lighten-4 primary--text"
-          >
-            <v-tab
-              v-for="item in items"
-              :key="item"
-              :ripple="false"
-              class="text-none"
+    <v-layout>
+      <v-card flat min-width="400px" color="background-lighten-4">
+        <v-system-bar color="primary" dark style="position: static">
+          <span> System bar </span>
+          <v-spacer></v-spacer>
+          <v-icon>mdi-minus</v-icon>
+          <v-icon>mdi-checkbox-blank-outline</v-icon>
+          <v-icon>mdi-close</v-icon>
+        </v-system-bar>
+        <v-toolbar flat color="primary-lighten-2" dark extension-height="36px">
+          <v-app-bar-nav-icon></v-app-bar-nav-icon>
+          <v-toolbar-title> Tool bar </v-toolbar-title>
+          <v-spacer></v-spacer>
+          <v-btn icon>
+            <v-icon>mdi-magnify</v-icon>
+          </v-btn>
+          <template v-slot:extension>
+            <v-tabs
+              v-model="tab"
+              align-with-title
+              height="36px"
+              hide-slider
+              active-class="bg-background-lighten-4 text-primary"
             >
+              <v-tab
+                v-for="item in items"
+                :key="item"
+                :value="item"
+                :ripple="false"
+                class="text-none"
+              >
+                {{ item }}
+              </v-tab>
+            </v-tabs>
+          </template>
+        </v-toolbar>
+        <v-window v-model="tab">
+          <v-window-item
+            v-for="item in items"
+            :key="item"
+            :value="item"
+            :transition="false"
+            :reverse-transition="false"
+          >
+            <v-card-text>
+              <pre>{{ item }}</pre>
+            </v-card-text>
+          </v-window-item>
+        </v-window>
+      </v-card>
+    </v-layout>
+    <v-layout>
+      <v-card flat min-width="400px" color="background-lighten-4">
+        <v-system-bar color="primary" dark style="position: static;">
+          <span> System bar </span>
+          <v-spacer></v-spacer>
+          <v-icon>mdi-minus</v-icon>
+          <v-icon>mdi-checkbox-blank-outline</v-icon>
+          <v-icon>mdi-close</v-icon>
+        </v-system-bar>
+        <v-tabs
+          v-model="tab"
+          align-with-title
+          height="36px"
+          hide-slider
+          dark
+          background-color="primary-lighten-2"
+          active-class="bg-background-lighten-4 text-primary"
+        >
+          <v-tab
+            v-for="item in items"
+            :key="item"
+            :value="item"
+            :ripple="false"
+            class="text-none"
+          >
+            <v-badge content="1" color="error" inline>
               {{ item }}
-            </v-tab>
-          </v-tabs>
-        </template>
-      </v-toolbar>
-      <v-tabs-items v-model="tab">
-        <v-tab-item
-          v-for="item in items"
-          :key="item"
-          :transition="false"
-          :reverse-transition="false"
-        >
-          <v-card-text>
-            <pre>{{ item }}</pre>
-          </v-card-text>
-        </v-tab-item>
-      </v-tabs-items>
-    </v-card>
-    <v-card flat min-width="400px" color="background lighten-4">
-      <v-system-bar color="primary" dark>
-        <span> System bar </span>
-        <v-spacer></v-spacer>
-        <v-icon>mdi-minus</v-icon>
-        <v-icon>mdi-checkbox-blank-outline</v-icon>
-        <v-icon>mdi-close</v-icon>
-      </v-system-bar>
-      <v-tabs
-        v-model="tab"
-        align-with-title
-        height="36px"
-        hide-slider
-        dark
-        background-color="primary lighten-2"
-        active-class="background lighten-4 primary--text"
-      >
-        <v-tab
-          v-for="item in items"
-          :key="item"
-          :ripple="false"
-          class="text-none"
-        >
-          <v-badge content="1" color="error" inline>
-            {{ item }}
-          </v-badge>
-        </v-tab>
-      </v-tabs>
-      <v-tabs-items v-model="tab">
-        <v-tab-item
-          v-for="item in items"
-          :key="item"
-          :transition="false"
-          :reverse-transition="false"
-        >
-          <v-card-text>
-            <pre>{{ item }}</pre>
-          </v-card-text>
-        </v-tab-item>
-      </v-tabs-items>
-    </v-card>
+            </v-badge>
+          </v-tab>
+        </v-tabs>
+        <v-window v-model="tab">
+          <v-window-item
+            v-for="item in items"
+            :key="item"
+            :value="item"
+            :transition="false"
+            :reverse-transition="false"
+          >
+            <v-card-text>
+              <pre>{{ item }}</pre>
+            </v-card-text>
+          </v-window-item>
+        </v-window>
+      </v-card>
+    </v-layout>
   </div>
 </template>
 
