@@ -1,9 +1,7 @@
 import { defineConfig, loadEnv } from "vite";
-import vue from "@vitejs/plugin-vue2";
+import vue from "@vitejs/plugin-vue";
 import loadVersion from "vite-plugin-package-version";
 import graphql from "@rollup/plugin-graphql";
-import { VuetifyResolver } from "unplugin-vue-components/resolvers";
-import Components from "unplugin-vue-components/vite";
 import monacoEditorPlugin from "vite-plugin-monaco-editor";
 import path from "path-browserify";
 
@@ -14,7 +12,15 @@ export default ({ mode }) => {
   return defineConfig({
     server: { port: 8081 },
     plugins: [
-      vue(),
+      vue({
+        template: {
+          compilerOptions: {
+            compatConfig: {
+              MODE: 3,
+            },
+          },
+        },
+      }),
       loadVersion(),
       // @ts-ignore
       graphql(),
@@ -22,15 +28,15 @@ export default ({ mode }) => {
         languageWorkers: ["editorWorkerService"],
         publicPath: "monacoeditorwork", // default
       }),
-      Components({
-        resolvers: [VuetifyResolver()],
-      }),
     ],
     base: process.env.VITE_PUBLIC_PATH,
     resolve: {
       alias: {
         "@": path.resolve(__dirname, "src"),
         path: "path-browserify",
+        alias: {
+          vue: "@vue/compat",
+        },
       },
     },
   });
